refactor(core): 收敛运行时边界并整理项目结构
