+++ internal/publicartifacts/store.go
@@ -1,31 +1,15 @@
 package publicartifacts
 
 import (
-	"archive/tar"
-	"compress/gzip"
-	"crypto/hmac"
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"image"
-	_ "image/gif"
-	_ "image/jpeg"
-	_ "image/png"
-	"io"
-	"mime"
-	"net/http"
 	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
-	"unicode/utf8"
-
-	protocol "github.com/uvwt/agentdock-protocol"
 )
 
 const (
@@ -235,583 +219,4 @@ func normalizeNow(now time.Time) time.Time {
 func (s Store) EnsureSecret() error {
 	_, err := s.ensureSecret()
 	return err
-}
-
-func (s Store) ServeHTTP(w http.ResponseWriter, r *http.Request, prefix string) {
-	if r.Method != http.MethodGet && r.Method != http.MethodHead {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	id, name, ok := parsePublicPath(r.URL.Path, prefix)
-	if !ok {
-		http.NotFound(w, r)
-		return
-	}
-	expires, err := strconv.ParseInt(r.URL.Query().Get("expires"), 10, 64)
-	if err != nil || expires <= 0 {
-		http.NotFound(w, r)
-		return
-	}
-	if time.Now().UTC().Unix() > expires {
-		http.Error(w, http.StatusText(http.StatusGone), http.StatusGone)
-		return
-	}
-	meta, payload, err := s.resolveArtifact(id)
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	if meta.Filename != name || meta.ExpiresAt.Unix() != expires {
-		http.NotFound(w, r)
-		return
-	}
-	secret, err := s.ensureSecret()
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	expected := sign(secret, meta.ArtifactID, meta.Filename, expires, meta.SHA256)
-	if !hmac.Equal([]byte(expected), []byte(r.URL.Query().Get("sig"))) {
-		http.NotFound(w, r)
-		return
-	}
-	file, err := os.OpenInRoot(s.Root, payload)
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	defer file.Close()
-	info, err := file.Stat()
-	if err != nil || !info.Mode().IsRegular() || info.Size() != meta.Size {
-		http.NotFound(w, r)
-		return
-	}
-	payloadSHA, err := streamSHA256(file)
-	if err != nil || payloadSHA != meta.SHA256 {
-		http.NotFound(w, r)
-		return
-	}
-	if _, err := file.Seek(0, io.SeekStart); err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	w.Header().Set("Content-Type", firstNonEmpty(meta.MimeType, "application/octet-stream"))
-	w.Header().Set("Content-Disposition", mime.FormatMediaType(contentDisposition(meta.MimeType), map[string]string{"filename": meta.Filename}))
-	w.Header().Set("Cache-Control", "private, no-store")
-	w.Header().Set("X-Content-Type-Options", "nosniff")
-	// 即使未来某种主动内容被错误标成 inline，浏览器也不能执行脚本、提交表单或加载外部资源。
-	w.Header().Set("Content-Security-Policy", "sandbox; default-src 'none'; base-uri 'none'; form-action 'none'; frame-ancestors 'none'")
-	w.Header().Set("Referrer-Policy", "no-referrer")
-	http.ServeContent(w, r, meta.Filename, meta.CreatedAt, file)
-}
-
-func (s Store) Read(artifactID string, maxBytes int64) (Metadata, []byte, error) {
-	meta, payload, err := s.resolveArtifact(artifactID)
-	if err != nil {
-		return Metadata{}, nil, fmt.Errorf("read artifact metadata: %w", err)
-	}
-	if time.Now().UTC().After(meta.ExpiresAt) {
-		return Metadata{}, nil, errors.New("artifact has expired")
-	}
-	if maxBytes > 0 && meta.Size > maxBytes {
-		return Metadata{}, nil, fmt.Errorf("artifact size %d exceeds limit %d", meta.Size, maxBytes)
-	}
-	file, err := os.OpenInRoot(s.Root, payload)
-	if err != nil {
-		return Metadata{}, nil, fmt.Errorf("read artifact payload: %w", err)
-	}
-	defer file.Close()
-	data, err := io.ReadAll(file)
-	if err != nil {
-		return Metadata{}, nil, fmt.Errorf("read artifact payload: %w", err)
-	}
-	if int64(len(data)) != meta.Size {
-		return Metadata{}, nil, errors.New("artifact payload size does not match metadata")
-	}
-	digest := sha256.Sum256(data)
-	if hex.EncodeToString(digest[:]) != meta.SHA256 {
-		return Metadata{}, nil, errors.New("artifact payload checksum does not match metadata")
-	}
-	return meta, data, nil
-}
-
-// ReadChunk returns one verified slice of an Artifact for the outbound-only
-// Nexus Bridge. A zero maxBytes performs a metadata-only read for HTTP HEAD.
-func (s Store) ReadChunk(artifactID string, offset int64, maxBytes int) (Metadata, []byte, bool, error) {
-	if offset < 0 {
-		return Metadata{}, nil, false, errors.New("artifact offset must not be negative")
-	}
-	if maxBytes < 0 {
-		return Metadata{}, nil, false, errors.New("artifact chunk size must not be negative")
-	}
-	if maxBytes > protocol.MaxArtifactChunkBytes {
-		maxBytes = protocol.MaxArtifactChunkBytes
-	}
-	meta, payload, err := s.resolveArtifact(artifactID)
-	if err != nil {
-		return Metadata{}, nil, false, fmt.Errorf("read artifact metadata: %w", err)
-	}
-	if time.Now().UTC().After(meta.ExpiresAt) {
-		return Metadata{}, nil, false, errors.New("artifact has expired")
-	}
-	if offset > meta.Size {
-		return Metadata{}, nil, false, fmt.Errorf("artifact offset %d exceeds payload size %d", offset, meta.Size)
-	}
-
-	file, err := os.OpenInRoot(s.Root, payload)
-	if err != nil {
-		return Metadata{}, nil, false, fmt.Errorf("open artifact payload: %w", err)
-	}
-	defer file.Close()
-	info, err := file.Stat()
-	if err != nil || !info.Mode().IsRegular() || info.Size() != meta.Size {
-		return Metadata{}, nil, false, errors.New("artifact payload size does not match metadata")
-	}
-	if maxBytes == 0 {
-		return meta, nil, offset == meta.Size, nil
-	}
-	// The first chunk verifies the immutable snapshot before any bytes leave the
-	// node. Nexus additionally verifies the complete stream against this digest.
-	if offset == 0 {
-		payloadSHA, hashErr := streamSHA256(file)
-		if hashErr != nil || payloadSHA != meta.SHA256 {
-			return Metadata{}, nil, false, errors.New("artifact payload checksum does not match metadata")
-		}
-	}
-	if _, err := file.Seek(offset, io.SeekStart); err != nil {
-		return Metadata{}, nil, false, fmt.Errorf("seek artifact payload: %w", err)
-	}
-	remaining := meta.Size - offset
-	readBytes := int64(maxBytes)
-	if remaining < readBytes {
-		readBytes = remaining
-	}
-	data, err := io.ReadAll(io.LimitReader(file, readBytes))
-	if err != nil {
-		return Metadata{}, nil, false, fmt.Errorf("read artifact chunk: %w", err)
-	}
-	if int64(len(data)) != readBytes {
-		return Metadata{}, nil, false, errors.New("artifact payload changed while it was being read")
-	}
-	nextOffset := offset + int64(len(data))
-	return meta, data, nextOffset == meta.Size, nil
-}
-
-func (s Store) Cleanup(now time.Time) error {
-	if now.IsZero() {
-		now = time.Now().UTC()
-	} else {
-		now = now.UTC()
-	}
-	if err := os.MkdirAll(s.Root, 0o700); err != nil {
-		return fmt.Errorf("create public artifacts root: %w", err)
-	}
-	entries, err := os.ReadDir(s.Root)
-	if err != nil {
-		return err
-	}
-	var cleanupErrs []error
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			continue
-		}
-		dir := filepath.Join(s.Root, entry.Name())
-		info, statErr := entry.Info()
-		if statErr != nil {
-			cleanupErrs = append(cleanupErrs, fmt.Errorf("inspect artifact directory %s: %w", dir, statErr))
-			continue
-		}
-		meta, metaErr := readMetadataPath(filepath.Join(dir, "metadata.json"))
-		payloadInfo, payloadErr := os.Stat(filepath.Join(dir, "payload"))
-		oldBroken := now.Sub(info.ModTime()) > 24*time.Hour
-		remove := metaErr == nil && meta.ExpiresAt.Before(now) ||
-			metaErr != nil && oldBroken ||
-			payloadErr != nil && oldBroken ||
-			payloadErr == nil && !payloadInfo.Mode().IsRegular() && oldBroken
-		if remove {
-			if err := os.RemoveAll(dir); err != nil {
-				cleanupErrs = append(cleanupErrs, fmt.Errorf("remove artifact directory %s: %w", dir, err))
-			}
-		}
-	}
-	return errors.Join(cleanupErrs...)
-}
-
-func (s Store) resolveArtifact(id string) (Metadata, string, error) {
-	// Artifact IDs are generated as 16 random bytes encoded as lowercase hex.
-	// IsLocal plus the hex whitelist rejects absolute and traversing identifiers;
-	// os.Root below additionally prevents symlink traversal from escaping Root.
-	if !filepath.IsLocal(id) || !validArtifactID(id) {
-		return Metadata{}, "", errors.New("invalid artifact id")
-	}
-	root, err := os.OpenRoot(s.Root)
-	if err != nil {
-		return Metadata{}, "", err
-	}
-	defer root.Close()
-	metadataPath := filepath.Join(id, "metadata.json")
-	data, err := root.ReadFile(metadataPath)
-	if err != nil {
-		return Metadata{}, "", err
-	}
-	meta, err := decodeMetadata(data)
-	if err != nil {
-		return Metadata{}, "", err
-	}
-	if meta.ArtifactID != id {
-		return Metadata{}, "", errors.New("artifact metadata id does not match path")
-	}
-	return meta, filepath.Join(id, "payload"), nil
-}
-
-func validArtifactID(id string) bool {
-	if len(id) != 32 {
-		return false
-	}
-	for i := 0; i < len(id); i++ {
-		if (id[i] < '0' || id[i] > '9') && (id[i] < 'a' || id[i] > 'f') {
-			return false
-		}
-	}
-	return true
-}
-
-func readMetadataPath(path string) (Metadata, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return Metadata{}, err
-	}
-	return decodeMetadata(data)
-}
-
-func decodeMetadata(data []byte) (Metadata, error) {
-	var meta Metadata
-	if err := json.Unmarshal(data, &meta); err != nil {
-		return Metadata{}, err
-	}
-	if meta.ArtifactID == "" || meta.Filename == "" || meta.SHA256 == "" || meta.Size < 0 || meta.ExpiresAt.IsZero() {
-		return Metadata{}, errors.New("invalid metadata")
-	}
-	return meta, nil
-}
-
-func (s Store) ensureSecret() ([]byte, error) {
-	secret, err := readSecret(s.SecretPath)
-	if err == nil {
-		return secret, nil
-	}
-	if !os.IsNotExist(err) {
-		return nil, err
-	}
-
-	secretDir := filepath.Dir(s.SecretPath)
-	if err := os.MkdirAll(secretDir, 0o700); err != nil {
-		return nil, fmt.Errorf("create secret dir: %w", err)
-	}
-	if err := os.Chmod(secretDir, 0o700); err != nil {
-		return nil, fmt.Errorf("secure secret dir: %w", err)
-	}
-	secret = make([]byte, 32)
-	if _, err := rand.Read(secret); err != nil {
-		return nil, fmt.Errorf("generate public url secret: %w", err)
-	}
-
-	// 先完整写入同目录临时文件，再用硬链接以“不覆盖”语义发布。
-	// 并发进程只有一个能创建最终路径，其他进程读取胜出的同一份密钥。
-	tmp, err := os.CreateTemp(secretDir, ".public-url-secret-*")
-	if err != nil {
-		return nil, fmt.Errorf("create public url secret temp file: %w", err)
-	}
-	tmpPath := tmp.Name()
-	defer os.Remove(tmpPath)
-	if err := tmp.Chmod(0o600); err != nil {
-		_ = tmp.Close()
-		return nil, fmt.Errorf("secure public url secret temp file: %w", err)
-	}
-	if _, err := tmp.WriteString(hex.EncodeToString(secret) + "\n"); err != nil {
-		_ = tmp.Close()
-		return nil, fmt.Errorf("write public url secret temp file: %w", err)
-	}
-	if err := tmp.Sync(); err != nil {
-		_ = tmp.Close()
-		return nil, fmt.Errorf("sync public url secret temp file: %w", err)
-	}
-	if err := tmp.Close(); err != nil {
-		return nil, fmt.Errorf("close public url secret temp file: %w", err)
-	}
-	if err := os.Link(tmpPath, s.SecretPath); err == nil {
-		return secret, nil
-	} else if !os.IsExist(err) {
-		return nil, fmt.Errorf("publish public url secret: %w", err)
-	}
-	return readSecret(s.SecretPath)
-}
-
-func readSecret(path string) ([]byte, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	decoded, err := hex.DecodeString(strings.TrimSpace(string(data)))
-	if err != nil || len(decoded) < 32 {
-		return nil, errors.New("public url secret is invalid")
-	}
-	if err := os.Chmod(path, 0o600); err != nil {
-		return nil, fmt.Errorf("secure public url secret: %w", err)
-	}
-	return decoded, nil
-}
-
-func parsePublicPath(pathValue, prefix string) (string, string, bool) {
-	rest := strings.TrimPrefix(pathValue, prefix)
-	if rest == pathValue || rest == "" {
-		return "", "", false
-	}
-	parts := strings.Split(rest, "/")
-	if len(parts) != 2 {
-		return "", "", false
-	}
-	id, err1 := url.PathUnescape(parts[0])
-	name, err2 := url.PathUnescape(parts[1])
-	if err1 != nil || err2 != nil || id == "" || name == "" || id != filepath.Base(id) || name != filepath.Base(name) || strings.Contains(name, "\\") {
-		return "", "", false
-	}
-	return id, name, true
-}
-
-func sign(secret []byte, id, filename string, expires int64, sha string) string {
-	mac := hmac.New(sha256.New, secret)
-	_, _ = fmt.Fprintf(mac, "%s\n%s\n%d\n%s", id, filename, expires, sha)
-	return hex.EncodeToString(mac.Sum(nil))
-}
-
-func retention(seconds int) time.Duration {
-	if seconds <= 0 {
-		return DefaultRetention
-	}
-	d := time.Duration(seconds) * time.Second
-	if d > MaxRetention {
-		return MaxRetention
-	}
-	return d
-}
-
-func randomHex(byteCount int) (string, error) {
-	if byteCount <= 0 {
-		return "", errors.New("random byte count must be positive")
-	}
-	b := make([]byte, byteCount)
-	if _, err := rand.Read(b); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(b), nil
-}
-
-func copyFile(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return fmt.Errorf("open publish source: %w", err)
-	}
-	defer in.Close()
-	out, err := os.OpenFile(dst, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o600)
-	if err != nil {
-		return fmt.Errorf("create payload: %w", err)
-	}
-	_, copyErr := io.Copy(out, in)
-	closeErr := out.Close()
-	if copyErr != nil {
-		return fmt.Errorf("copy payload: %w", copyErr)
-	}
-	if closeErr != nil {
-		return fmt.Errorf("close payload: %w", closeErr)
-	}
-	return nil
-}
-
-func writeTarGz(src, dst string) error {
-	out, err := os.OpenFile(dst, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o600)
-	if err != nil {
-		return fmt.Errorf("create archive payload: %w", err)
-	}
-	gz := gzip.NewWriter(out)
-	tw := tar.NewWriter(gz)
-	walkErr := filepath.WalkDir(src, func(path string, d os.DirEntry, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		info, err := d.Info()
-		if err != nil {
-			return err
-		}
-		name, err := filepath.Rel(filepath.Dir(src), path)
-		if err != nil {
-			return err
-		}
-		name = filepath.ToSlash(name)
-		if name == "." {
-			return nil
-		}
-		h, err := tar.FileInfoHeader(info, "")
-		if err != nil {
-			return err
-		}
-		h.Name = name
-		if err := tw.WriteHeader(h); err != nil {
-			return err
-		}
-		if info.Mode().IsRegular() {
-			in, err := os.Open(path)
-			if err != nil {
-				return err
-			}
-			_, err = io.Copy(tw, in)
-			closeErr := in.Close()
-			if err != nil {
-				return err
-			}
-			if closeErr != nil {
-				return closeErr
-			}
-		}
-		return nil
-	})
-	closeTar := tw.Close()
-	closeGz := gz.Close()
-	closeOut := out.Close()
-	if walkErr != nil {
-		return fmt.Errorf("archive source: %w", walkErr)
-	}
-	if closeTar != nil {
-		return closeTar
-	}
-	if closeGz != nil {
-		return closeGz
-	}
-	if closeOut != nil {
-		return closeOut
-	}
-	return nil
-}
-
-func fileSHA256(path string) (string, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return "", err
-	}
-	defer file.Close()
-	return streamSHA256(file)
-}
-
-func streamSHA256(reader io.Reader) (string, error) {
-	hash := sha256.New()
-	if _, err := io.Copy(hash, reader); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(hash.Sum(nil)), nil
-}
-
-func imageDimensions(path, mimeType string) (int, int) {
-	if !strings.HasPrefix(strings.ToLower(strings.TrimSpace(mimeType)), "image/") {
-		return 0, 0
-	}
-	file, err := os.Open(path)
-	if err != nil {
-		return 0, 0
-	}
-	defer file.Close()
-	cfg, _, err := image.DecodeConfig(file)
-	if err != nil {
-		return 0, 0
-	}
-	return cfg.Width, cfg.Height
-}
-
-func contentDisposition(mimeType string) string {
-	value, _, err := mime.ParseMediaType(strings.ToLower(strings.TrimSpace(mimeType)))
-	if err != nil {
-		value = strings.ToLower(strings.TrimSpace(mimeType))
-	}
-	switch value {
-	case "image/png", "image/jpeg", "image/gif", "image/webp", "image/avif", "image/bmp", "text/plain":
-		return "inline"
-	default:
-		// HTML、SVG、XML 等主动内容必须下载，不能在 AgentDock 同源下解释执行。
-		return "attachment"
-	}
-}
-
-func detectMime(path, filename string, archive bool) string {
-	if archive {
-		return "application/gzip"
-	}
-	if mt := mime.TypeByExtension(filepath.Ext(filename)); mt != "" {
-		return mt
-	}
-	f, err := os.Open(path)
-	if err != nil {
-		return "application/octet-stream"
-	}
-	defer f.Close()
-	buf := make([]byte, 512)
-	n, err := f.Read(buf)
-	if err != nil && !errors.Is(err, io.EOF) {
-		return "application/octet-stream"
-	}
-	if n == 0 {
-		return "application/octet-stream"
-	}
-	return http.DetectContentType(buf[:n])
-}
-
-func safeDownloadName(value string) string {
-	value = strings.ToValidUTF8(value, "_")
-	value = strings.Map(func(char rune) rune {
-		if char < 0x20 || char == 0x7f {
-			return '_'
-		}
-		return char
-	}, value)
-	value = strings.ReplaceAll(strings.TrimSpace(value), "\\", "/")
-	value = filepath.Base(value)
-	if value == "" || value == "." || value == ".." || strings.ContainsAny(value, "/\\") {
-		return "artifact.bin"
-	}
-	if len(value) <= 240 {
-		return value
-	}
-
-	// 下载名受 HTTP 头和文件系统共同约束。按字节限制时必须停在 rune 边界，
-	// 否则长中文名会生成非法 UTF-8，并继续污染 metadata、URL 与响应头。
-	ext := filepath.Ext(value)
-	if len(ext) >= 240 {
-		ext = ""
-	}
-	base := strings.TrimSuffix(value, ext)
-	base = truncateUTF8Bytes(base, 240-len(ext))
-	if base == "" {
-		return "artifact.bin"
-	}
-	return base + ext
-}
-
-func truncateUTF8Bytes(value string, maxBytes int) string {
-	if maxBytes <= 0 {
-		return ""
-	}
-	if len(value) <= maxBytes {
-		return value
-	}
-	for maxBytes > 0 && !utf8.ValidString(value[:maxBytes]) {
-		maxBytes--
-	}
-	return value[:maxBytes]
-}
-
-func firstNonEmpty(values ...string) string {
-	for _, v := range values {
-		if strings.TrimSpace(v) != "" {
-			return strings.TrimSpace(v)
-		}
-	}
-	return ""
 }
