+++ internal/tool/recall/helpers.go
@@ -1,10 +1,6 @@
 package recall
 
-import (
-	"io"
-
-	toolcore "github.com/uvwt/agentdock/internal/tool/core"
-)
+import toolcore "github.com/uvwt/agentdock/internal/tool/core"
 
 type Result = toolcore.Result
 type ToolError = toolcore.ToolError
@@ -17,7 +13,4 @@ func toolErrorDetails(code, message, category string, details map[string]any) *T
 }
 func toolErrorCause(code, message, category string, details map[string]any, cause error) *ToolError {
 	return toolcore.NewErrorCause(code, message, category, details, cause)
-}
-func readBoundedBody(reader io.Reader, maxBytes int64) ([]byte, error) {
-	return toolcore.ReadBoundedBody(reader, maxBytes)
 }
