+++ internal/tool/task/helpers.go
@@ -1,8 +1,6 @@
 package task
 
 import (
-	"io"
-
 	"github.com/uvwt/agentdock/internal/textutil"
 	toolcore "github.com/uvwt/agentdock/internal/tool/core"
 )
@@ -15,9 +13,6 @@ func toolErrorDetails(code, message, category string, details map[string]any) *T
 }
 func toolErrorCause(code, message, category string, details map[string]any, cause error) *ToolError {
 	return toolcore.NewErrorCause(code, message, category, details, cause)
-}
-func readBoundedBody(reader io.Reader, maxBytes int64) ([]byte, error) {
-	return toolcore.ReadBoundedBody(reader, maxBytes)
 }
 func truncateString(value string, maxBytes int) string {
 	return textutil.SafeTruncateString(value, maxBytes).Text
