+++ scripts/test/install_linux_test.go
@@ -1,0 +1,153 @@
+package scripts
+
+import (
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"testing"
+)
+
+func TestInstallLinuxRemovesLegacyNexusCredentials(t *testing.T) {
+	data, err := os.ReadFile("../install/install-linux-platform.sh")
+	if err != nil {
+		t.Fatalf("read install-linux-platform.sh: %v", err)
+	}
+	script := string(data)
+	forbidden := []string{
+		"local nexus_token=\"$7\"",
+		"printf 'AGENTDOCK_NEXUS_TOKEN=%s\\n' \"$nexus_token\"",
+		"NexusDock API 是否需要 token？",
+		"nexus_token=\"$(prompt_secret 'NexusDock token')\"",
+	}
+	for _, value := range forbidden {
+		if strings.Contains(script, value) {
+			t.Fatalf("install-linux-platform.sh still contains legacy Nexus credential handling: %s", value)
+		}
+	}
+	if !strings.Contains(script, "AGENTDOCK_NEXUS_ENDPOINT|AGENTDOCK_NEXUS_TOKEN") {
+		t.Fatal("install-linux-platform.sh must remove legacy Nexus credentials from an existing env file")
+	}
+	for _, removed := range []string{"AGENTDOCK_NEXUS_DEVICE_NAME", "AGENTDOCK_NEXUS_HEARTBEAT_SECONDS", "Nexus 设备名"} {
+		if strings.Contains(script, removed) {
+			t.Fatalf("install-linux-platform.sh still contains removed device-agent config %q", removed)
+		}
+	}
+}
+func TestLinuxInstallerIntegratesCloudflareTunnelWithoutLeakingToken(t *testing.T) {
+	data, err := os.ReadFile("../install/install-linux-platform.sh")
+	if err != nil {
+		t.Fatalf("read install-linux-platform.sh: %v", err)
+	}
+	script := string(data)
+	for _, want := range []string{
+		"你是否有已接入 Cloudflare 的域名？",
+		"AGENTDOCK_CLOUDFLARE_TUNNEL_TOKEN",
+		"AGENTDOCK_TUNNEL_MODE=$mode",
+		"TUNNEL_TOKEN=$token",
+		"EnvironmentFile=$cloudflared_env_file",
+		"service launch-core --runtime-root $runtime_root",
+		"tunnel launch --runtime-root $runtime_root",
+		"write_runtime_manifest",
+		`"service_manager": "$service_manager"`,
+		"AGENTDOCK_SERVER_URL=%s\\n",
+		"AGENTDOCK_OAUTH_ENABLED=%s\\n",
+		"AGENTDOCK_OAUTH_PASSWORD=%s\\n",
+		"AGENTDOCK_OAUTH_TOKEN_SECRET=%s\\n",
+		"Bearer Token、OAuth 均已启用",
+		`server_url="$TUNNEL_PUBLIC_URL"`,
+	} {
+		if !strings.Contains(script, want) {
+			t.Fatalf("install-linux-platform.sh missing Cloudflare Tunnel integration %q", want)
+		}
+	}
+	if strings.Contains(script, "--token $token") || strings.Contains(script, "--token \\$TUNNEL_TOKEN") {
+		t.Fatal("cloudflared token must be provided through its private environment file, not process arguments")
+	}
+	for _, legacy := range []string{
+		"ExecStart=$cloudflared_binary tunnel",
+		`command="$cloudflared_binary"`,
+		"ExecStart=$source_dir/bin/agentdock \\",
+	} {
+		if strings.Contains(script, legacy) {
+			t.Fatalf("installer still emits legacy direct runtime command %q", legacy)
+		}
+	}
+}
+func TestLinuxInstallerPreservesCredentialsAndCapturesQuickURL(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("Linux installer shell behavior is covered by the Alpine native runtime E2E")
+	}
+	tempDir := t.TempDir()
+	envFile := filepath.Join(tempDir, "agentdock.env")
+	initial := strings.Join([]string{
+		"AGENTDOCK_HOST=127.0.0.9",
+		"AGENTDOCK_PORT=19999",
+		"AGENTDOCK_AUTH_TOKEN=stable-token",
+		"AGENTDOCK_OAUTH_ENABLED=true",
+		"AGENTDOCK_OAUTH_PASSWORD=stable-oauth-password",
+		"AGENTDOCK_OAUTH_TOKEN_SECRET=stable-oauth-secret-0123456789abcdef",
+		"AGENTDOCK_BROWSER_ENABLED=true",
+		"",
+	}, "\n")
+	if err := os.WriteFile(envFile, []byte(initial), 0o600); err != nil {
+		t.Fatal(err)
+	}
+
+	script := `
+set -Eeuo pipefail
+source ../install/install-linux-platform.sh
+run_root() {
+  if [[ "$1" == systemctl ]]; then
+    return 0
+  fi
+  if [[ "$1" == install ]]; then
+    shift
+    local args=()
+    while (( $# > 0 )); do
+      case "$1" in
+        -o|-g) shift 2 ;;
+        *) args+=("$1"); shift ;;
+      esac
+    done
+    command install "${args[@]}"
+    return
+  fi
+  "$@"
+}
+write_env_file "$TEST_ENV_FILE" 127.0.0.1 8765 stable-token info \
+  https://new.trycloudflare.com yes true stable-oauth-password stable-oauth-secret-0123456789abcdef
+cloudflared_service_active() { return 0; }
+cloudflared_quick_url() { printf 'https://new.trycloudflare.com'; }
+start_cloudflared_service systemd agentdock-cloudflared quick ""
+printf '\nCAPTURED=%s\n' "$TUNNEL_PUBLIC_URL"
+`
+	cmd := exec.Command("bash", "-c", script)
+	cmd.Dir = "."
+	cmd.Env = append(os.Environ(), "TEST_ENV_FILE="+envFile)
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		t.Fatalf("run installer functions: %v\n%s", err, output)
+	}
+	got, err := os.ReadFile(envFile)
+	if err != nil {
+		t.Fatal(err)
+	}
+	text := string(got)
+	for _, want := range []string{
+		"AGENTDOCK_BROWSER_ENABLED=true",
+		"AGENTDOCK_AUTH_TOKEN=stable-token",
+		"AGENTDOCK_OAUTH_ENABLED=true",
+		"AGENTDOCK_OAUTH_PASSWORD=stable-oauth-password",
+		"AGENTDOCK_OAUTH_TOKEN_SECRET=stable-oauth-secret-0123456789abcdef",
+		"AGENTDOCK_SERVER_URL=https://new.trycloudflare.com",
+	} {
+		if !strings.Contains(text, want) {
+			t.Fatalf("rewritten env missing %q:\n%s", want, text)
+		}
+	}
+	if !strings.Contains(string(output), "CAPTURED=https://new.trycloudflare.com") {
+		t.Fatalf("quick URL was not captured: %s", output)
+	}
+}
